introduce linen-scene subproject
to share logic using linen-repository
