+++ linen-modern/src/main/scala/x7c1/linen/modern/display/settings/ChannelRowAdapter.scala
@@ -3,9 +3,9 @@ package x7c1.linen.modern.display.settings
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
 import x7c1.linen.glue.res.layout.{SettingMyChannelRow, SettingMyChannelRowFooter, SettingMyChannelRowItem}
-import x7c1.linen.modern.init.settings.preset.{MenuSelected, OnMenuSelectedListener}
 import x7c1.linen.repository.account.AccountIdentifiable
 import x7c1.linen.repository.channel.my.{MyChannel, MyChannelFooter, MyChannelRow}
+import x7c1.linen.scene.channel.menu.{MenuSelected, OnMenuSelectedListener}
 import x7c1.wheat.lore.resource.AdapterDelegatee
 import x7c1.wheat.modern.decorator.Imports._
