+++ linen-modern/src/main/scala/x7c1/linen/modern/init/dev/CreateRecordsDelegatee.scala
@@ -8,7 +8,8 @@ import x7c1.linen.database.control.{DatabaseHelper, LinenDatabase}
 import x7c1.linen.glue.res.layout.DevCreateRecordsLayout
 import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.glue.service.ServiceLabel.Updater
-import x7c1.linen.modern.init.updater.{UpdaterMethods, UpdaterServiceDelegatee}
+import x7c1.linen.scene.updater.UpdaterMethods
+import x7c1.linen.scene.updater.UpdaterServiceTypes.ActionTypeSample
 import x7c1.wheat.macros.intent.ServiceCaller
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.Imports._
@@ -29,7 +30,7 @@ class CreateRecordsDelegatee (
 
   def setup(): Unit = {
     getBroadcastManager.registerReceiver(
-      receiver, new IntentFilter(UpdaterServiceDelegatee.ActionTypeSample))
+      receiver, new IntentFilter(ActionTypeSample))
 
     layout.toolbar onClickNavigation { _ =>
       activity.finish()
