+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsSelectedFragment.scala
@@ -1,18 +1,12 @@
 package x7c1.linen.modern.init.settings.preset
 
-import android.app.Activity
 import android.os.Bundle
 import android.view.{LayoutInflater, View, ViewGroup}
 import x7c1.linen.glue.res.layout.{SettingPresetChannelRow, SettingPresetTabSelected}
-import x7c1.linen.glue.service.ServiceControl
-import x7c1.linen.glue.service.ServiceLabel.Updater
-import x7c1.linen.modern.init.updater.UpdaterMethods
 import x7c1.linen.repository.channel.preset.SelectedPresetChannelsAccessor
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.macros.fragment.TypedFragment
-import x7c1.wheat.macros.intent.ServiceCaller
 import x7c1.wheat.macros.logger.Log
-import x7c1.wheat.modern.menu.popup.{PopupMenuBox, PopupMenuItem}
 
 class ArgumentsForSelected(
   val accountId: Long,
@@ -47,22 +41,4 @@ class PresetsSelectedFragment extends TypedFragment[ArgumentsForSelected] with P
   }
 }
 
-class OnMenuForSelected(
-  activity: Activity with ServiceControl,
-  accountId: Long) extends OnMenuSelectedListener {
 
-  override def onMenuSelected(e: MenuSelected) = {
-    val items = Seq(
-      itemToLoadSources(e.channelId)
-    )
-    PopupMenuBox(activity, e.targetView, items).show()
-  }
-  private def itemToLoadSources(channelId: Long) =
-    PopupMenuItem("Load all sources"){ _ =>
-      Log info s"[init] channel:$channelId"
-      ServiceCaller.using[UpdaterMethods].
-        startService(activity, activity getClassOf Updater){
-          _ loadChannelSources (channelId, accountId)
-        }
-    }
-}
