+++ linen-scene/src/main/scala/x7c1/linen/scene/channel/menu/OnMenuSelectedListener.scala
@@ -1,0 +1,5 @@
+package x7c1.linen.scene.channel.menu
+
+trait OnMenuSelectedListener {
+  def onMenuSelected(e: MenuSelected): Unit
+}
