+++ linen-scene/src/main/scala/x7c1/linen/scene/updater/UpdaterServiceTypes.scala
@@ -1,0 +1,5 @@
+package x7c1.linen.scene.updater
+
+object UpdaterServiceTypes {
+  val ActionTypeSample = "hoge"
+}
